fix(observability-langfuse): isolate package-missing test into own file
vi.resetModules() alone did not fix the flake: the global beforeEach
in langfuse.test.ts unconditionally registers
vi.doMock('langfuse', { Langfuse: FakeLangfuse }), which races the
empty mock and still wins in CI (trace captured → "expected 1 to be
+0"). Moved the case to langfuse-missing.test.ts, which has no
FakeLangfuse mock at all, so there is no client to capture a trace and
the missing/broken-install path is deterministic. Asserts warn + no
throw instead of trace count. Coverage unchanged (lines 98.38%, gate
90%). Stable across repeated local runs (8/8, 2 files).
